requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.35
stripe==10.11.0
Werkzeug==3.0.4
WooCommerce==3.0.0
WTForms==3.1.2